requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
AWSIoTPythonSDK==1.4.7
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
entrypoints==0.3
Flask==1.1.1
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.3
luma.core==1.12.0
luma.oled==3.1.0
MarkupSafe==1.1.1
Pillow==5.4.0
pycrypto==2.6.1
pyftdi==0.40.6
PyGObject==3.30.4
pyserial==3.4
python-dotenv==0.10.3
pyusb==1.0.2
redis==3.3.11
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
smbus2==0.3.0
spidev==3.4
ssh-import-id==5.7
timeloop==1.0.2
urllib3==1.24.1
Werkzeug==0.16.0

